legacy/backend/requirements.txt
flask
flask-cors
kubernetes
flasgger
paramiko
pyyaml
